refactor: LottoMachine 파일 경로 controller로 이동
